Split session command IPC handlers
